# Chore databse optimizations