--- a/Gemfile
+++ b/Gemfile
@@ -40,3 +40,6 @@ group :development, :test do
   gem 'teaspoon-jasmine'
   gem 'selenium-webdriver'
 end
+
+
+gem "mumuki-domain", github: "mumuki/mumuki-domain", branch: "chore-database-optimization"