--- a/app/helpers/icons_helper.rb
+++ b/app/helpers/icons_helper.rb
@@ -7,9 +7,9 @@ module IconsHelper
     fa_icon name, options.merge(class: 'fa-fw fixed-icon')
   end
 
-  def exercise_status_icon(exercise)
-    link_to exercise_status_fa_icon(exercise),
-            exercise_path(exercise) if current_user?
+  def assignment_status_icon(assignment)
+    link_to contextualization_fa_icon(assignment),
+            exercise_path(assignment.exercise) if current_user?
   end
 
   def language_icon(language)
@@ -20,10 +20,6 @@ module IconsHelper
 
   def contextualization_fa_icon(contextualization)
     fa_icon(*icon_for(contextualization))
-  end
-
-  def exercise_status_fa_icon(exercise)
-    contextualization_fa_icon(exercise.assignment_for(current_user))
   end
 
   def discussion_status_fa_icon(discussion)