--- a/app/helpers/progress_bar_helper.rb
+++ b/app/helpers/progress_bar_helper.rb
@@ -1,8 +1,8 @@
 module ProgressBarHelper
   include IconsHelper
 
-  def class_for_progress_list_item(exercise, active)
-    "progress-list-item text-center #{class_for_exercise(exercise)} #{active ? 'active' : ''}"
+  def class_for_progress_list_item(assignment, active)
+    "progress-list-item text-center #{icon_class_for(assignment)} #{active ? 'active' : ''}"
   end
 
 end