--- a/app/views/chapters/show.html.erb
+++ b/app/views/chapters/show.html.erb
@@ -28,7 +28,7 @@
 
       <% @chapter.lessons.includes(guide: :exercises).each do |lesson| %>
         <h4><%= lesson.number %>. <%= link_to_path_element lesson, mode: :plain %></h4>
-        <%= render partial: 'layouts/progress_listing', locals: { exercises: lesson.exercises } %>
+        <%= render partial: 'layouts/progress_listing', locals: { guide: lesson.guide } %>
       <% end %>
     </div>
   <% end %>