--- a/app/views/exercises/show.html.erb
+++ b/app/views/exercises/show.html.erb
@@ -65,6 +65,6 @@
     <%= render partial: 'layouts/modals/level_up' %>
   <% end %>
 
-  <%= render partial: 'layouts/modals/guide_corollary', locals: {guide: @guide} %>
+  <%= render partial: 'layouts/modals/guide_corollary', locals: {guide: @guide} if @stats.almost_done? %>
   <%= render partial: 'layouts/modals/new_message', locals: {exercise: @exercise} if should_render_message_input?(@exercise) %>
 <% end if current_user? %>