--- a/app/views/layouts/_guide.html.erb
+++ b/app/views/layouts/_guide.html.erb
@@ -2,7 +2,7 @@
 
 <%= render partial: 'layouts/authoring', locals: {guide: @guide} %>
 
-<% if subject.timed? && subject.started?(current_user) && !current_user.teacher? %>
+<% if subject.timed? && @stats.started? && !current_user.teacher? %>
   <%= render partial: 'layouts/timer', locals: {end_time: subject.real_end_time(current_user)} %>
 <% end %>