--- a/app/views/layouts/_progress_bar.html.erb
+++ b/app/views/layouts/_progress_bar.html.erb
@@ -1,12 +1,14 @@
 <div class="progress-list-flex">
-  <% guide.exercises.each do |e| %>
+  <% assignments = guide.assignments_for(current_user) %>
+  <% assignments.each do |assignment| %>
+    <% exercise = assignment.exercise %>
     <a
-      <%= turbolinks_enable_for e %>
-      href="<%= exercise_path(e)%>"
-      aria-label="<%= e.navigable_name %>"
-      title="<%= e.navigable_name %>"
-      data-mu-exercise-id="<%= e.id %>"
-      class="<%= class_for_progress_list_item(e, e == actual)%>">
+      <%= turbolinks_enable_for exercise %>
+      href="<%= exercise_path(exercise)%>"
+      aria-label="<%= exercise.navigable_name %>"
+      title="<%= exercise.navigable_name %>"
+      data-mu-exercise-id="<%= exercise.id %>"
+      class="<%= class_for_progress_list_item(assignment, exercise == actual)%>">
     </a>
   <% end %>
 </div>