--- a/app/views/layouts/_progress_listing.html.erb
+++ b/app/views/layouts/_progress_listing.html.erb
@@ -1,9 +1,9 @@
 <ul class="progress-listing">
-  <% exercises.each do |exercise| %>
-      <% cache [exercise, exercise.status_for(current_user), Organization.current] do %>
-          <li <%= turbolinks_enable_for(exercise) %>>
-            <%= exercise_status_icon exercise %>
-            <%= link_to_path_element(exercise) %>
+  <% guide.assignments_for(current_user).each do |assignment| %>
+      <% cache [assignment.exercise, assignment.status, Organization.current] do %>
+          <li <%= turbolinks_enable_for(assignment.exercise) %>>
+            <%= assignment_status_icon assignment %>
+            <%= link_to_path_element(assignment.exercise) %>
           </li>
       <% end %>
   <% end %>