# Goal

To implement - and document - some database optimization that will impact on chapter, lessons and exercise page. 
This is part of a series of database and view optimizations, that will come after this initial PR. 
 
# :warning: Warnings

Depends on https://github.com/mumuki/mumuki-domain/pull/177

# :memo: Details

All commits are labeled with an optimization number that corresponds also to domain's PR. See https://github.com/mumuki/mumuki-domain/pull/177

## :page_facing_up:  Chapters Page

### :arrow_backward:  Before

* :clock1:  ActiveRecord Time: 327.7ms
* :hash: Number of queries: 141

```
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2251], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2252], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2254], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2255], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2257], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2745], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2747], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3251], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3255], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3258], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3259], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3868], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3873], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3877], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3892], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3901], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 84], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 87], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 89], ["LIMIT", 1]]
Assignment Load (0.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 95], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2737], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2738], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2739], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2740], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2743], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3876], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3880], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3881], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3886], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3888], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3893], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3902], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 86], ["LIMIT", 1]]
Assignment Load (0.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 92], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2256], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2734], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3247], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3252], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3253], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3257], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3885], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3891], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 85], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 91], ["LIMIT", 1]]
Assignment Load (0.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 96], ["LIMIT", 1]]
Assignment Load (0.8ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2732], ["LIMIT", 1]]
Assignment Load (0.8ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2748], ["LIMIT", 1]]
Assignment Load (0.8ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3870], ["LIMIT", 1]]
Assignment Load (0.8ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3889], ["LIMIT", 1]]
Assignment Load (0.8ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3890], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2248], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2741], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2746], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3249], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3256], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 90], ["LIMIT", 1]]
Assignment Load (0.9ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 94], ["LIMIT", 1]]
Assignment Load (1.0ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3874], ["LIMIT", 1]]
Assignment Load (1.0ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3896], ["LIMIT", 1]]
Assignment Load (1.1ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3882], ["LIMIT", 1]]
Assignment Load (1.1ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3883], ["LIMIT", 1]]
Assignment Load (1.2ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 88], ["LIMIT", 1]]
Assignment Load (1.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2249], ["LIMIT", 1]]
Assignment Load (1.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3878], ["LIMIT", 1]]
Assignment Load (1.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3895], ["LIMIT", 1]]
Assignment Load (1.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3898], ["LIMIT", 1]]
Assignment Load (1.4ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2735], ["LIMIT", 1]]
Assignment Load (1.4ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3250], ["LIMIT", 1]]
Assignment Load (1.4ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3884], ["LIMIT", 1]]
Assignment Load (1.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3254], ["LIMIT", 1]]
Assignment Load (1.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3897], ["LIMIT", 1]]
Assignment Load (1.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3899], ["LIMIT", 1]]
Assignment Load (1.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3900], ["LIMIT", 1]]
Assignment Load (1.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 93], ["LIMIT", 1]]
Assignment Load (1.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2247], ["LIMIT", 1]]
Assignment Load (1.7ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3879], ["LIMIT", 1]]
Assignment Load (10.0ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3903], ["LIMIT", 1]]
Assignment Load (14.2ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3887], ["LIMIT", 1]]
Assignment Load (2.0ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2250], ["LIMIT", 1]]
Assignment Load (2.1ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2736], ["LIMIT", 1]]
Assignment Load (2.1ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2742], ["LIMIT", 1]]
Assignment Load (2.2ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2253], ["LIMIT", 1]]
Assignment Load (2.2ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3872], ["LIMIT", 1]]
Assignment Load (2.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2744], ["LIMIT", 1]]
Assignment Load (2.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3248], ["LIMIT", 1]]
Assignment Load (2.3ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3869], ["LIMIT", 1]]
Assignment Load (2.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3894], ["LIMIT", 1]]
Assignment Load (2.6ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 2733], ["LIMIT", 1]]
Assignment Load (4.0ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3875], ["LIMIT", 1]]
Assignment Load (5.5ms)  SELECT  "assignments".* FROM "assignments" WHERE "assignments"."exercise_id" = $1 AND "assignments"."submitter_id" = 8 LIMIT $2  [["exercise_id", 3871], ["LIMIT", 1]]
Avatar Load (0.4ms)  SELECT  "avatars".* FROM "avatars" WHERE "avatars"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
Avatar Load (0.4ms)  SELECT  "avatars".* FROM "avatars" WHERE "avatars"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
Avatar Load (0.8ms)  SELECT  "avatars".* FROM "avatars" WHERE "avatars"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
Book Load (0.4ms)  SELECT  "books".* FROM "books" WHERE "books"."id" = $1 LIMIT $2  [["id", 31], ["LIMIT", 1]]
Chapter Load (0.5ms)  SELECT  "chapters".* FROM "chapters" WHERE "chapters"."id" = $1 ORDER BY "chapters"."number" ASC LIMIT $2  [["id", 93], ["LIMIT", 1]]
Chapter Load (1.9ms)  SELECT  "chapters".* FROM "chapters" WHERE "chapters"."id" = $1 ORDER BY "chapters"."number" ASC LIMIT $2  [["id", 93], ["LIMIT", 1]]
Discussion Load (1.3ms)  SELECT "discussions".* FROM "discussions" INNER JOIN "subscriptions" ON "discussions"."id" = "subscriptions"."discussion_id" WHERE "subscriptions"."user_id" = $1 ORDER BY "discussions"."created_at" DESC  [["user_id", 8]]
Exercise Load (0.6ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 427]]
Exercise Load (0.8ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 425]]
Exercise Load (0.9ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 15]]
Exercise Load (0.9ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 260]]
Exercise Load (0.9ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 426]]
Exercise Load (1.3ms)  SELECT "exercises".* FROM "exercises" WHERE "exercises"."guide_id" = $1 ORDER BY "exercises"."number" ASC  [["guide_id", 307]]
…