@felipecalvo 2021-01-25
This is awesome so far :heart: 

Not querying each assignment individually was long overdue :grimacing::grimacing: